Trait ksched::prique::PriqueTrait

pub trait PriqueTrait<T> {
    const P: usize;

    fn new() -> Self
    where
        Self: Sized;

    fn push(&mut self, nice: usize, t: T);

    fn pop(&mut self) -> Option<(usize, T)>;

    fn len(&self) -> usize;

    fn check(&self);

    fn max_nice() -> usize { ... }
}

Priority queue with eight priority levels.

Associated Constants

const P: usize

Max nice value.

Required methods

fn new() -> Self where
Self: Sized,

Create a new priority queue. Panic if oom.

fn push(&mut self, nice: usize, t: T)

Push back a item with specific priority.

Panic if nice value is greater or equal to P. Nice 0 is of highest priority.

fn pop(&mut self) -> Option<(usize, T)>

Pop out the item of largest priority(lowest nice value).

Items of same priority will be poped in order of FIFO.

fn len(&self) -> usize

Get the number of elements in queue.

fn check(&self)

Checker funtion for debug.

Provided methods

fn max_nice() -> usize

Return the max nice value.