Module ksched::sync

Synchronization primitives.

Re-exports

pub use super::mpsc;

Structs

Condvar
A Condition Variable

Mutex
An async mutex.

MutexGuard
A guard that releases the mutex when dropped.

RwLock
An async reader-writer lock.

RwLockReadGuard
A guard that releases the read lock when dropped.

RwLockUpgradableReadGuard
A guard that releases the upgradable read lock when dropped.

RwLockWriteGuard
A guard that releases the write lock when dropped.

Spinlock
This type provides MUTual EXclusion based on spinning.

SpinlockGuard
A guard to which the protected data can be accessed